Extract class Interfaces from Application
PR-URL: https://github.com/metarhia/impress/issues/1534
