+++ lib/application.js
@@ -1,9 +1,10 @@
 'use strict';
 
 const { node, npm, metarhia } = require('./dependencies.js');
-const { path, events, fsp } = node;
-const { metautil, metavm, metawatch } = metarhia;
-const { Procedure } = require('./procedure.js');
+const { path, events } = node;
+const { metavm, metawatch } = metarhia;
+
+const { Interfaces } = require('./interfaces.js');
 const { Cache } = require('./cache.js');
 const { Resources } = require('./resources.js');
 
@@ -16,29 +17,19 @@ class Error extends global.Error {
 
 const SANDBOX = { ...metavm.COMMON_CONTEXT, Error, node, npm, metarhia };
 
-const getSignature = (method) => {
-  const src = method.toString();
-  const signature = metautil.between(src, '({', '})');
-  if (signature === '') return [];
-  return signature.split(',').map((s) => s.trim());
-};
-
 class Application extends events.EventEmitter {
   constructor() {
     super();
     this.initialization = true;
     this.finalization = false;
-
     this.root = process.cwd();
     this.path = path.join(this.root, 'application');
-    this.apiPath = this.absolute('api');
 
-    this.api = {};
-    this.signatures = {};
-    this.static = new Resources(this.absolute('static'));
-    this.resources = new Resources(this.absolute('resources'));
-    this.lib = new Cache(this.absolute('lib'));
-    this.domain = new Cache(this.absolute('domain'));
+    this.api = new Interfaces('api', this);
+    this.static = new Resources('static', this);
+    this.resources = new Resources('resources', this);
+    this.lib = new Cache('lib', this);
+    this.domain = new Cache('domain', this);
     this.starts = [];
 
     this.Application = Application;
@@ -61,7 +52,7 @@ class Application extends events.EventEmitter {
     await Promise.allSettled([
       this.static.load(),
       this.resources.load(),
-      this.loadPlace('api', this.apiPath),
+      this.api.load(),
       (async () => {
         await this.lib.load();
         await this.domain.load();
@@ -102,22 +93,6 @@ class Application extends events.EventEmitter {
     this.sandbox = metavm.createContext(sandbox);
   }
 
-  async createScript(fileName) {
-    try {
-      const code = await fsp.readFile(fileName, 'utf8');
-      if (!code) return null;
-      const src = 'context => ' + code;
-      const options = { context: this.sandbox };
-      const { exports } = new metavm.MetaScript(fileName, src, options);
-      return exports;
-    } catch (err) {
-      if (err.code !== 'ENOENT') {
-        this.console.error(err.stack);
-      }
-      return null;
-    }
-  }
-
   getMethod(iname, ver, methodName) {
     const iface = this.api[iname];
     if (!iface) return null;
@@ -128,72 +103,10 @@ class Application extends events.EventEmitter {
     return proc;
   }
 
-  async loadMethod(fileName) {
-    const rel = fileName.substring(this.apiPath.length + 1);
-    if (!rel.includes(path.sep)) return;
-    const [interfaceName, methodFile] = rel.split(path.sep);
-    if (!methodFile.endsWith('.js')) return;
-    const name = path.basename(methodFile, '.js');
-    const [iname, ver] = interfaceName.split('.');
-    const version = parseInt(ver, 10);
-    const script = await this.createScript(fileName);
-    if (!script) return;
-    const proc = new Procedure(script, this);
-    let iface = this.api[iname];
-    const { api } = this.sandbox;
-    let internalInterface = api[iname];
-    if (!iface) {
-      this.api[iname] = iface = { default: version };
-      api[iname] = internalInterface = {};
-    }
-    if (version > iface.default) iface.default = version;
-    let methods = iface[ver];
-    if (!methods) iface[ver] = methods = {};
-    const { method } = proc;
-    methods[name] = proc;
-    internalInterface[name] = method;
-    this.cacheSignature(iname + '.' + version, name, method);
-  }
-
-  async removeMethod(fileName) {
-    const rel = fileName.substring(this.apiPath.length + 1);
-    if (!rel.includes(path.sep)) return;
-    const [interfaceName, methodFile] = rel.split(path.sep);
-    if (!methodFile.endsWith('.js')) return;
-    const name = path.basename(methodFile, '.js');
-    const [iname, ver] = interfaceName.split('.');
-    const version = parseInt(ver, 10);
-    const iface = this.api[iname];
-    if (!iface) return null;
-    const methods = iface[version.toString()];
-    if (methods) delete methods[name];
-    const internalInterface = this.sandbox.api[iname];
-    if (internalInterface) delete internalInterface[name];
-  }
-
-  cacheSignature(interfaceName, methodName, method) {
-    let interfaceMethods = this.signatures[interfaceName];
-    if (!interfaceMethods) {
-      this.signatures[interfaceName] = interfaceMethods = {};
-    }
-    interfaceMethods[methodName] = getSignature(method);
-  }
-
   execute(method) {
     return method().catch((err) => {
       this.console.error(err.stack);
     });
-  }
-
-  async loadPlace(place, placePath) {
-    const files = await fsp.readdir(placePath, { withFileTypes: true });
-    for (const file of files) {
-      if (file.name.startsWith('.')) continue;
-      const filePath = path.join(placePath, file.name);
-      if (file.isDirectory()) await this.loadPlace(place, filePath);
-      else if (place === 'api') await this.loadMethod(filePath);
-    }
-    this.watcher.watch(placePath);
   }
 
   startWatch() {
@@ -216,22 +129,14 @@ class Application extends events.EventEmitter {
       if (node.worker.threadId === 1) {
         this.console.debug('Reload: /' + relPath);
       }
-      if (place === 'api') this.loadMethod(filePath);
-      else if (place === 'static') this.static.change(filePath);
-      else if (place === 'resources') this.resources.change(filePath);
-      else if (place === 'lib') await this.lib.change(filePath);
-      else if (place === 'domain') await this.domain.change(filePath);
+      this[place].change(filePath);
     });
 
     this.watcher.on('delete', async (filePath) => {
       const relPath = filePath.substring(this.path.length + 1);
       const sepIndex = relPath.indexOf(path.sep);
       const place = relPath.substring(0, sepIndex);
-      if (place === 'api') this.removeMethod(filePath);
-      else if (place === 'static') this.static.delete(filePath);
-      else if (place === 'resources') this.resources.delete(filePath);
-      else if (place === 'lib') await this.lib.delete(filePath);
-      else if (place === 'domain') await this.domain.delete(filePath);
+      this[place].delete(filePath);
       if (node.worker.threadId === 1) {
         this.console.debug('Deleted: /' + relPath);
       }
@@ -245,7 +150,7 @@ class Application extends events.EventEmitter {
       const iface = this.api[iname];
       if (!iface) continue;
       const version = ver === '*' ? iface.default : parseInt(ver);
-      intro[iname] = this.signatures[iname + '.' + version];
+      intro[iname] = this.api.signatures[iname + '.' + version];
     }
     return intro;
   }
